Fixed a crash in gupnp_didl_lite_contributor_get_name
The crash occurs if the DIDL contains an empty artists tag, e.g.,
<upnp:artist></upnp:artist>.  A call to gupnp_didl_lite_object_get_artists
results in a list of one element.  However, invoking
gupnp_didl_lite_contributor_get_name on this element causes a crash.
The fix works by modifying get_contributor_list_by_name so that it does
not add empty artists nodes to the list of contributors it creates.

https://bugzilla.gnome.org/show_bug.cgi?id=674319