Link to gobject-2.0 and libxml-2.0 as well
These where previously pulled in by underlying libraries
which leaked these dependencies. It happens to be that gupnp-av
needs these too and should not rely on anything else pulling it in.

(For static linking to work, these should probably be added to
the pkg-config files Requires.private field as well.)

https://bugzilla.gnome.org/show_bug.cgi?id=685477